requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
banks==2.1.2
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.26.165
botocore==1.29.165
bs4==0.0.2
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
cloudinary==1.37.0
colorama==0.4.6
contourpy==1.3.0
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distlib==0.3.9
distro==1.9.0
Django==4.1.13
django-autoslug==1.9.8
django-pagedown==2.2.1
django-storages==1.13.2
djongo==1.3.6
dnspython==2.7.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fake-useragent==1.5.1
feedparser==6.0.11
filelock==3.17.0
filetype==1.2.0
fonttools==4.55.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.1
griffe==1.7.3
h11==0.16.0
hf-xet==1.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
llama-cloud==0.1.19
llama-cloud-services==0.6.21
llama-index-core==0.12.34.post1
llama-index-embeddings-huggingface==0.5.3
llama-index-embeddings-openai==0.3.1
llama-index-llms-huggingface==0.5.0
llama-index-llms-openai==0.3.38
llama-index-readers-file==0.4.7
llama-parse==0.6.21
lxml==5.3.0
Markdown==3.6
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.4.3
multitasking==0.0.11
mypy_extensions==1.1.0
narwhals==1.37.1
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
openai==1.77.0
packaging==24.2
pandas==2.2.3
parse==1.20.2
peewee==3.17.8
pillow==11.0.0
pipenv==2022.6.7
platformdirs==4.3.7
plotly==5.24.1
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pyee==11.1.1
pymongo==4.10.1
pyparsing==3.2.0
pypdf==5.4.0
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-html==0.10.0
rpds-py==0.24.0
s3transfer==0.6.2
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.2.4
streamlit==1.45.0
striprtf==0.0.26
sympy==1.14.0
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
virtualenv==20.29.1
virtualenv-clone==0.5.7
w3lib==2.2.1
watchdog==6.0.0
webencodings==0.5.1
websockets==10.4
wrapt==1.17.2
yahoo-fin==0.8.9.1
yarl==1.20.0
yfinance==0.2.50
zipp==3.21.0
